+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Better error message when comparing different enum types in the filter language
 - Include descriptions for scalars and directives in introspection and schema output
 - Fix issue using the default `SchemaBuilderOptions` with a mutation method that has a list argument. The argument input type was not being correctly added to the schema. Not an issue if you were already adding the type manually
+- #336 - Fix conversion of non-nullable/nullable types
 
 ## Changes