+++ CHANGELOG.md
@@ -1,9 +1,17 @@
+# 5.1.1
+
+## Fixes
+
+- Fix #337 - Description attribute accidentally updating name instead of description
+
 # 5.1.0
 
 ## Changes
+
 - Upgrade to the latest standard Antlr4 - the parser/tool used for the filter expression strings. Fixing precedence of operators
 
 ## Fixes
+
 - #319 - Only convert strings to `Guid` or `DateTime` when required
 - If you add a field connecting back to the Query Context that ends in a method to go from a list to a single result the graph object projection will now correctly be inserted before the last call. Example
 
@@ -23,12 +31,16 @@ schema.Query().AddField("externalData", "Get Data")
     .ResolveWithService<ExternalDataService>((p, srv) => srv.GetData());
 ```
 
-In the above case if you select 
+In the above case if you select
 
 ```gql
 {
   externalData {
-    movie { director { name } }
+    movie {
+      director {
+        name
+      }
+    }
   }
 }
 ```
@@ -48,6 +60,7 @@ db.Movies.Where(m => m.Id == ed.Id)
 # 5.0.1
 
 ## Fixes
+
 - Fix #314 - Some clean up of the Antlr4 grammer for the filter expressions
 
 # 5.0.0
@@ -55,9 +68,11 @@ db.Movies.Where(m => m.Id == ed.Id)
 Make sure to check out the changes 5.0.0-beta1
 
 ## Breaking Changes
+
 - Generated schema type name for field sort inputs now include the name of the schema type the field is on to avoid conflicts
 
 ## Changes
+
 - `IField.AddExtension` now returns the `IField`
 - `UseSort()` field extension now can take a list of default sort fields e.g.
 - `Broadcaster` (inbuilt `IObservable<TType>` you can use for subscriptions) now has a `OnUnsubscribe` callback